add mem cache
