+++ package.json
@@ -7,6 +7,7 @@
     "fastify": "^2.4.1",
     "fastify-static": "^2.4.0",
     "gh-badges": "^2.1.0",
+    "mem": "^5.0.0",
     "package-compliant": "^1.1.1",
     "package-json": "^6.3.0"
   },
